+++ tests/test_parser.py
@@ -10,7 +10,6 @@ import textwrap
 
 import pytest
 
-from glean_parser import lint
 from glean_parser import metrics
 from glean_parser import parser
 
@@ -1040,63 +1039,3 @@ def test_no_internal_fields_exposed():
         indent=2,
     )
     assert expected_json == out_json
-
-
-def test_unit_not_accepted_on_nonquantity():
-    results = parser.parse_objects(
-        [
-            util.add_required(
-                {
-                    "category": {"metric": {"type": "counter", "unit": "quantillions"}},
-                }
-            ),
-        ]
-    )
-    errs = list(results)
-    assert len(errs) == 1
-    assert "got an unexpected keyword argument 'unit'" in str(errs[0])
-
-
-def test_unit_accepted_on_custom_dist():
-    results = parser.parse_objects(
-        [
-            util.add_required(
-                {
-                    "category": {
-                        "metric": {
-                            "type": "custom_distribution",
-                            "unit": "quantillions",
-                            "range_max": 100,
-                            "bucket_count": 100,
-                            "histogram_type": "linear",
-                        }
-                    },
-                }
-            ),
-        ]
-    )
-    errs = list(results)
-    assert len(errs) == 0
-
-
-def test_unit_accepted_on_string_but_warns():
-    results = parser.parse_objects(
-        [
-            util.add_required(
-                {
-                    "category": {
-                        "metric": {
-                            "type": "string",
-                            "unit": "quantillions",
-                        }
-                    },
-                }
-            ),
-        ]
-    )
-    errs = list(results)
-    assert len(errs) == 0
-
-    nits = lint.lint_metrics(results.value)
-    assert len(nits) == 1
-    assert set(["UNIT_ON_STRING"]) == set(v.check_name for v in nits)