+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Warn about empty or TODO-tagged data reviews in the list ([#634](https://github.com/mozilla/glean_parser/pull/634))
+- Allow `unit` field on all metrics, but warn for all but quantity and custom distribution ([#636](https://github.com/mozilla/glean_parser/pull/636))
 
 ## 10.0.2